pyproject.toml
[tool.poetry]
name = "chimera"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.7"
djangorestframework = "^3.13.1"
wagtail = "^2.16.1"
django-cors-headers = "^3.11.0"
psycopg2 = "^2.9.3"
requests = "^2.27.1"
Pillow = "^9.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
